refactor: 拆分空闲状态组件到不同模块
