+++ src/renderer/components/components/DynamicIslandStateContent.tsx
@@ -26,7 +26,7 @@
 
 import type { JSX } from 'react';
 import type { NotificationData, WeatherData } from '../../store/types';
-import { IdleContent } from '../states/idle/IdleContent';
+import { IdleContent } from '../states/idle';
 import { HoverContent } from '../states/hover/HoverContent';
 import { NotificationContent } from '../states/notification/NotificationContent';
 import { ExpandedContent } from '../states/expand/ExpandedContent';
