+++ src/renderer/components/states/idle/index.ts
@@ -1,0 +1,1 @@
+export { IdleContent } from './IdleContent';
