+++ src/renderer/components/states/idle/utils/padZero.ts
@@ -1,0 +1,3 @@
+export function padZero(value: number): string {
+  return value < 10 ? `0${value}` : `${value}`;
+}
