+++ src/renderer/components/states/index.ts
@@ -24,7 +24,7 @@
  * @author 鸡哥
  */
 
-export { IdleContent } from './idle/IdleContent';
+export { IdleContent } from './idle';
 export { HoverContent } from './hover/HoverContent';
 export { ExpandedContent } from './expand/ExpandedContent';
 export { MaxExpandContent } from './maxExpand/MaxExpandContent';
